requirements.txt
streamlit
opencv-python
easyocr
pillow
numpy
google-generativeai
